+++ four/src/iam/resource/saml_provider.rs
@@ -1,12 +1,11 @@
-use four_derive::ManagedResource;
-use nutype::nutype;
-use serde::Serialize;
-
-use crate::core::{
-    function::{HaveAtt, RefInner, Referenced},
-    service::IAM,
-    Arn, LogicalId, Tag,
+use crate::{
+    core::{
+        function::{HaveAtt, RefInner, Referenced},
+        LogicalId, Tag,
+    },
+    iam::{SAMLMetadataDocument, SAMLProviderArn, SAMLProviderName},
 };
+use four_derive::ManagedResource;
 
 #[derive(ManagedResource, Clone)]
 #[resource_type = "AWS::IAM::SAMLProvider"]
@@ -15,27 +14,6 @@ pub struct SAMLProvider {
     name: Option<SAMLProviderName>,
     saml_metadata_document: Option<SAMLMetadataDocument>,
     tags: Option<Vec<Tag>>,
-}
-
-#[nutype(
-    validate(not_empty, len_char_max = 128, regex = r#"[\w._-]+"#),
-    derive(Debug, Clone, Serialize)
-)]
-pub struct SAMLProviderName(String);
-
-#[nutype(
-    validate(len_char_min = 1000, len_char_max = 10000000),
-    derive(Debug, Clone, Serialize)
-)]
-pub struct SAMLMetadataDocument(String);
-
-#[derive(Debug, Clone, Serialize)]
-pub struct SAMLProviderArn(Arn<IAM>);
-
-impl From<Arn<IAM>> for SAMLProviderArn {
-    fn from(value: Arn<IAM>) -> Self {
-        SAMLProviderArn(value)
-    }
 }
 
 impl Referenced for SAMLProvider {
