+++ four/src/iam/resource/server_certificate.rs
@@ -1,15 +1,14 @@
-use four_derive::ManagedResource;
-use nutype::nutype;
-use serde::Serialize;
-
 use crate::{
     core::{
         function::{HaveAtt, RefInner, Referenced},
-        service::IAM,
-        Arn, LogicalId, Tag,
+        LogicalId, Tag,
     },
-    iam::path::Path,
+    iam::{
+        CertificateBody, CertificateChain, Path, PrivateKey, ServerCertificateArn,
+        ServerCertificateName,
+    },
 };
+use four_derive::ManagedResource;
 
 #[derive(ManagedResource, Clone)]
 #[resource_type = "AWS::IAM::ServerCertificate"]
@@ -20,51 +19,6 @@ pub struct ServerCertificate {
     path: Option<Path>,
     private_key: Option<PrivateKey>,
     tags: Option<Vec<Tag>>,
-}
-
-#[nutype(
-    validate(
-        not_empty,
-        len_char_max = 16384,
-        regex = r#"[\u0009\u000A\u000D\u0020-\u00FF]+"#
-    ),
-    derive(Debug, Clone, Serialize)
-)]
-pub struct CertificateBody(String);
-
-#[nutype(
-    validate(
-        not_empty,
-        len_char_max = 2097152,
-        regex = r#"[\u0009\u000A\u000D\u0020-\u00FF]+"#
-    ),
-    derive(Debug, Clone, Serialize)
-)]
-pub struct CertificateChain(String);
-
-#[nutype(
-    validate(
-        not_empty,
-        len_char_max = 16384,
-        regex = r#"[\u0009\u000A\u000D\u0020-\u00FF]+"#
-    ),
-    derive(Debug, Clone, Serialize)
-)]
-pub struct PrivateKey(String);
-
-#[nutype(
-    validate(not_empty, len_char_max = 128, regex = r#"[\w+=,.@-]+"#),
-    derive(Debug, Clone, Serialize)
-)]
-pub struct ServerCertificateName(String);
-
-#[derive(Debug, Clone, Serialize)]
-pub struct ServerCertificateArn(Arn<IAM>);
-
-impl From<Arn<IAM>> for ServerCertificateArn {
-    fn from(value: Arn<IAM>) -> Self {
-        ServerCertificateArn(value)
-    }
 }
 
 impl Referenced for ServerCertificate {
